COMBINADO con cambios de Fran

+++ Nominas/Gestion_Nomina.cs
@@ -10,9 +10,9 @@ namespace Nominas
     {
       
         //TODO: Cargar ajustes por dfecto
-        private static short horas = 0; //Total de horas trabajadas en la semana (se almacena en el temporal)
-        private static short extra = 0; //Horas extra trabajadas en la semana (se almacena en el temporal)
-        private static short precio = 0; //Precio de la hora trabajada (se almacena en el temporal)
+        private static int horas = 0; //Total de horas trabajadas en la semana (se almacena en el temporal)
+        private static int extra = 0; //Horas extra trabajadas en la semana (se almacena en el temporal)
+        private static float precio = 0; //Precio de la hora trabajada (se almacena en el temporal)
         private static int jornada = 0; //Horas a partir de las cuales se consideran extras (se almacena en el temporal)
         private static float horasExtra = 0.0F;//Multplicador de precio para horas extra (se almacena en el temporal)
 
@@ -41,7 +41,7 @@ namespace Nominas
         //Fichero temporal lleva las horas por semana
 
         //Cálculo de las horas extra
-        private static short CalculoExtra(short hora, int jornada)
+        private static int CalculoExtra(int hora, int jornada)
         {
             extra = (short)(hora - jornada);
             return extra;
@@ -49,7 +49,7 @@ namespace Nominas
         }
 
         //Cálculo del salario bruto
-        private static float CalculoSalarioBruto(short horas, int jornada, float precio)
+        private static float CalculoSalarioBruto(int horas, int jornada, float precio)
         {
             if (horas > jornada)
             {
@@ -65,7 +65,7 @@ namespace Nominas
         }
 
         //Calculo del salario Extra
-        private static float CalculoSalarioExtra(short extra, float precio)
+        private static float CalculoSalarioExtra(int extra, float precio)
         {
             return (extra * precio * horasExtra);
         }
